refactor: dto folders
